api/requirements.txt
mysql-connector-python~=8.1.0
Flask~=2.3.2
requests~=2.31.0
